cmake: add cmake Find<package>.cmake for thrift
In some cases, when using jaeger-cpp-client, we use thrift provided from
distro packages, these does not contain thriftConfig.cmake.
For such scenarios, we pivot to do not make find_package CONFIG mode as
default, instead we use basic signature[0]. We make use of
`CMAKE_MODULE_PATH` to search for FindThrift.cmake, which will find and
set all thrift::thrift's target properties.

This appends cmake/modules dir to CMAKE_MODULE_PATH, which is used
as all modules requiring find_package mode with basic_signature.

[0] https://cmake.org/cmake/help/latest/command/find_package.html
The command arguments determine which of the above modes is used. When
the basic signature is used, the command searches in Module mode first.
If the package is not found, the search falls back to Config mode.

Signed-off-by: Deepika Upadhyay <dupadhya@redhat.com>