resource sdk: Update Resource::Merge function docs
The documented behaviour of Resource::Merge is the opposite of what it
actually does.

As per https://cplusplus.com/reference/unordered_map/unordered_map/insert/,
when calling `insert` on an unordered_map, each element is inserted only
if its key is not equivalent to the key of any other element already in
the container.

Signed-off-by: ricekot <github@ricekot.com>