Allow cross-compilation with protobuf
The upstream config module from protobuf always declares targets
for the protobuf library and protoc of the same platform. This means
one of the two is always wrong, as protoc needs to be run on the
build machine, but the library must be linked for the "host" platform.

A small change in the CMake scaffold allows the user to pass
`-DPROTOBUF_PROTOC_EXECUTABLE=<path/to/protoc>` to cmake when creating
the build tree.