Implementing configurable aggregation cardinality limit and collector emitting all views instead of last view
Issues

Cardinality of the Aggregation limit (2000) is hardcoded and no ability to configure as part of the meter/aggregation
Meter stores meter-name to storage in storage_registry_. When multiple views are added, the last view overrides the previous view, and collector can only emit the last view's metrics that are added.

Fixes
Introducing cardinality as configurable parameter in AggregationConfig, and implementing the limit from AggregationConfig in Storage
Changing the storage_registry_ to keep track of view_name to storage, instead of metric_name to storage. So when multiple views are added, different view-names will be collected by collector.
Please provide a brief description of the changes here.