[Metrics] Allow registering one callback for multiple instruments
This commit adds a pair of methods `Meter::RegisterCallback` and
`Meter::DeregisterCallback` which allow you to set up one callback
which can record observations for multiple observable instruments.

The callback will be passed an object of type MultiObserverResult, which
can be used to get an ObserverResultT for an instrument.

This allows a single expensive operation to yield multiple different
observable metrics without the need to call said expensive operation
multiple times.